% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branding.R
\name{contato_rodape_tag}
\alias{contato_rodape_tag}
\title{Linha de contato do rodape (usa pndr_contato)}
\usage{
contato_rodape_tag()
}
\description{
Linha de contato do rodape (usa pndr_contato)
}
\keyword{internal}
