% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branding.R
\name{header_marca}
\alias{header_marca}
\title{Header do painel com a marca configuravel (usa pndr_titulo/subtitulo/logo)}
\usage{
header_marca()
}
\description{
Header do painel com a marca configuravel (usa pndr_titulo/subtitulo/logo)
}
\keyword{internal}
