% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_mdata}
\alias{painel_mdata}
\title{Metadados dos indicadores (mdata) com rotulo para selectize}
\usage{
painel_mdata(con)
}
\description{
Metadados dos indicadores (mdata) com rotulo para selectize
}
\keyword{internal}
