% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_valores}
\alias{painel_valores}
\title{Valores de um indicador (todas as localidades e refdates)}
\usage{
painel_valores(con, mdata_id)
}
\description{
Valores de um indicador (todas as localidades e refdates)
}
\keyword{internal}
